[refactor] #5 snackbar 확장함수로 변경 및 util 패키지로 이동
