+++ app/src/main/java/com/yhj/gdscandroidstudy/ui/home/HomeFragment.kt
@@ -5,8 +5,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.yhj.gdscandroidstudy.collectWhenStarted
 import com.yhj.gdscandroidstudy.databinding.FragmentHomeBinding
+import com.yhj.gdscandroidstudy.util.collectWhenStarted
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class HomeFragment : Fragment() {
