+++ app/src/main/java/com/yhj/gdscandroidstudy/ui/mypage/MyPageFragment.kt
@@ -6,9 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.yhj.gdscandroidstudy.collectWhenStarted
 import com.yhj.gdscandroidstudy.databinding.FragmentMypageBinding
 import com.yhj.gdscandroidstudy.ui.edit.EditActivity
+import com.yhj.gdscandroidstudy.util.collectWhenStarted
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MyPageFragment : Fragment() {
