+++ app/src/main/java/com/yhj/gdscandroidstudy/util/Extensions.kt
@@ -1,9 +1,11 @@
-package com.yhj.gdscandroidstudy
+package com.yhj.gdscandroidstudy.util
 
+import android.view.View
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
+import com.google.android.material.snackbar.Snackbar
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 
@@ -15,4 +17,8 @@ fun <T> LifecycleOwner.collectWhenStarted(flow: Flow<T>, block: suspend (T) -> U
             }
         }
     }
+}
+
+fun View.showSnackBar(msg: String) {
+    Snackbar.make(this, msg, Snackbar.LENGTH_SHORT).show()
 }
